Fix an issue with chat field focus commands.
Occasionally, hitting '/' when the chat field was not focused would
cause two slashes to appear in the chat field. Similarly, hitting '.'
when the field was not focused could cause a period to appear at the end
of the chat field.

To fix this issue we prevent the event from performing its default
action. In addition, we stop propagation of the event (just to be safe).